Part writing in file
Create a new interface ISymptomWriter for writing symptoms in a file.
Create a new class WriteSymptomsInFile which implement this interface.
Remove code writing in file from the main to WriteSymptomsInFile.
Changes in iSymptomReader : use a Treemap instead of a list to avoid
duplications and to order symptoms in alphabetical order.
Change getSymptoms method in ReadSymptomDataFromFile to implement
correctly the new interface.